# Add new `Performance/UriDefaultParser` cop