--- a/config/enabled.yml
+++ b/config/enabled.yml
@@ -1601,6 +1601,10 @@ Performance/UnfreezeString:
   Description: 'Use unary plus to get an unfrozen string literal.'
   Enabled: true
 
+Performance/UriDefaultParser:
+  Description: 'Use `URI::DEFAULT_PARSER` instead of `URI::Parser.new`.'
+  Enabled: true
+
 #################### Rails #################################
 
 Rails/ActionFilter: