--- a/manual/cops_performance.md
+++ b/manual/cops_performance.md
@@ -786,3 +786,22 @@ String.new('something')
 +'something'
 +''
 ```
+
+## Performance/UriDefaultParser
+
+Enabled by default | Supports autocorrection
+--- | ---
+Enabled | Yes
+
+This cop identifies places where `URI::Parser.new`
+can be replaced by `URI::DEFAULT_PARSER`.
+
+### Example
+
+```ruby
+# bad
+URI::Parser.new
+
+# good
+URI::DEFAULT_PARSER
+```