This cop is inspired by https://github.com/bbatsov/rubocop/pull/4694#discussion_r136707061.

## Feature

This cop identifies places where `URI::Parser.new` can be replaced by `URI::DEFAULT_PARSER`.

```console
% cat /tmp/test.rb
# frozen_string_literal: true

require 'uri'

URI::Parser.new.make_regexp
```

```console
% bundle exec rubocop /tmp/test.rb
Inspecting 1 file
C

Offenses:

/tmp/test.rb:5:1: C: Use URI::DEFAULT_PARSER instead of URI::Parser.new.
URI::Parser.new.make_regexp
^^^^^^^^^^^^^^^

1 file inspected, 1 offense detected
```

## Target Problem

The following is a benchmark script.

```ruby
require 'benchmark/ips'
require 'uri'

Benchmark.ips do |x|
  x.report('URI::Parser.new')     { URI::Parser.new }
  x.report('URI::DEFAULT_PARSER') { URI::DEFAULT_PARSER }
  x.compare!
end
```

It is the result of running the above benchmark script.

```console
% ruby bench.rb
Warming up --------------------------------------
     URI::Parser.new    70.000  i/100ms
 URI::DEFAULT_PARSER   315.505k i/100ms
Calculating -------------------------------------
     URI::Parser.new    714.537  (± 4.2%) i/s -      3.570k in   5.005495s
 URI::DEFAULT_PARSER     10.950M (± 6.2%) i/s -     54.582M in   5.007247s

Comparison:
 URI::DEFAULT_PARSER: 10949722.8 i/s
     URI::Parser.new:      714.5 i/s - 15324.22x  slower
```

It is a constant as follows and it is fast because it does not instantiate every time.
https://github.com/ruby/ruby/blob/v2_4_1/lib/uri/common.rb#L22

-----------------

Before submitting the PR make sure the following are checked:

* [x] Wrote [good commit messages][1].
* [ ] Commit message starts with `[Fix #issue-number]` (if the related issue exists).
* [x] Used the same coding conventions as the rest of the project.
* [x] Feature branch is up-to-date with `master` (if not - rebase it).
* [x] Squashed related commits together.
* [x] Added tests.
* [x] Added an entry to the [Changelog](../blob/master/CHANGELOG.md) if the new code introduces user-observable changes. See [changelog entry format](../blob/master/CONTRIBUTING.md#changelog-entry-format).
* [x] All tests(`rake spec`) are passing.
* [x] The new code doesn't generate RuboCop offenses that are checked by `rake internal_investigation`.
* [x] The PR relates to *only* one subject with a clear title
  and description in grammatically correct, complete sentences.
* [x] Updated cop documentation with `rake generate_cops_documentation` (required only when you've added a new cop or changed the configuration/documentation of an existing cop).

[1]: http://chris.beams.io/posts/git-commit/